create use collaborative editor mode
